# frontend/shared/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatMessageDto": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "insufficient_context": {
            "title": "Insufficient Context",
            "type": "boolean"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ChatSourceDto"
            },
            "title": "Sources",
            "type": "array"
          },
          "trace_events": {
            "items": {
              "$ref": "#/components/schemas/ChatTraceEventDto"
            },
            "title": "Trace Events",
            "type": "array"
          }
        },
        "required": [
          "id",
          "role",
          "content",
          "sequence",
          "insufficient_context",
          "created_at",
          "sources",
          "trace_events"
        ],
        "title": "ChatMessageDto",
        "type": "object"
      },
      "ChatRunCreateRequest": {
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "limit": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "specialist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialist"
          }
        },
        "required": [
          "question"
        ],
        "title": "ChatRunCreateRequest",
        "type": "object"
      },
      "ChatRunDto": {
        "properties": {
          "conversation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConversationDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trace_events": {
            "items": {
              "$ref": "#/components/schemas/ChatTraceEventDto"
            },
            "title": "Trace Events",
            "type": "array"
          }
        },
        "required": [
          "id",
          "conversation_id",
          "status",
          "question",
          "error",
          "started_at",
          "finished_at",
          "trace_events",
          "conversation"
        ],
        "title": "ChatRunDto",
        "type": "object"
      },
      "ChatRunResponse": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "run": {
            "$ref": "#/components/schemas/ChatRunDto"
          }
        },
        "required": [
          "run"
        ],
        "title": "ChatRunResponse",
        "type": "object"
      },
      "ChatRunStartResponse": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "run": {
            "$ref": "#/components/schemas/ChatRunStartedDto"
          }
        },
        "required": [
          "run"
        ],
        "title": "ChatRunStartResponse",
        "type": "object"
      },
      "ChatRunStartedDto": {
        "properties": {
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "conversation_id",
          "run_id"
        ],
        "title": "ChatRunStartedDto",
        "type": "object"
      },
      "ChatSourceDto": {
        "properties": {
          "char_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char End"
          },
          "char_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char Start"
          },
          "chunk_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Index"
          },
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "document_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Name"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "qdrant_point_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qdrant Point Id"
          },
          "relative_raw_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative Raw Path"
          },
          "retrieval_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Query"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "used": {
            "default": false,
            "title": "Used",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "source_id",
          "source",
          "content_hash",
          "document_id",
          "document_name",
          "relative_raw_path",
          "qdrant_point_id",
          "chunk_index",
          "char_start",
          "char_end",
          "retrieval_query",
          "score",
          "excerpt"
        ],
        "title": "ChatSourceDto",
        "type": "object"
      },
      "ChatTraceEventDto": {
        "properties": {
          "agent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Name"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tool_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Name"
          }
        },
        "required": [
          "sequence",
          "event_type",
          "phase",
          "title",
          "status",
          "agent_name",
          "tool_name",
          "payload",
          "duration_ms"
        ],
        "title": "ChatTraceEventDto",
        "type": "object"
      },
...
      "ConversationDto": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageDto"
            },
            "title": "Messages",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "created_at",
          "updated_at",
          "messages"
        ],
        "title": "ConversationDto",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PipelineDocumentDto": {
        "properties": {
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "document_name": {
            "title": "Document Name",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "relative_raw_path": {
            "title": "Relative Raw Path",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "position",
          "document_name",
          "relative_raw_path",
          "status",
          "current_step",
          "error"
        ],
        "title": "PipelineDocumentDto",
        "type": "object"
      },
      "PipelineEventDto": {
        "properties": {
          "counters": {
            "additionalProperties": true,
            "title": "Counters",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "step": {
            "title": "Step",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "sequence",
          "timestamp",
          "step",
          "status",
          "message",
          "counters",
          "result"
        ],
        "title": "PipelineEventDto",
        "type": "object"
      },
      "PipelineRunCreateRequest": {
        "properties": {
          "document_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Document Ids",
            "type": "array"
          }
        },
        "title": "PipelineRunCreateRequest",
        "type": "object"
      },
      "PipelineRunDto": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/PipelineDocumentDto"
            },
            "title": "Documents",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/PipelineEventDto"
            },
            "title": "Events",
            "type": "array"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "summary",
          "error",
          "started_at",
          "finished_at",
          "created_at",
          "updated_at",
          "documents",
          "events"
        ],
        "title": "PipelineRunDto",
        "type": "object"
      },
      "PipelineRunListResponse": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/PipelineRunDto"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "runs"
        ],
        "title": "PipelineRunListResponse",
        "type": "object"
      },
      "PipelineRunResponse": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "run": {
            "$ref": "#/components/schemas/PipelineRunDto"
          }
        },
        "required": [
          "run"
        ],
        "title": "PipelineRunResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Medic RAG Dashboard",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/chat/runs": {
      "post": {
        "operationId": "start_chat_run_api_chat_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatRunStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Chat Run"
      }
    },
    "/api/chat/runs/{run_id}": {
      "get": {
        "operationId": "get_chat_run_api_chat_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chat Run"
      }
    },
    "/api/chat/runs/{run_id}/events": {
      "get": {
        "operationId": "stream_chat_run_events_api_chat_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Chat Run Events"
      }
    },
...
    "/api/pipeline-runs": {
      "get": {
        "operationId": "list_pipeline_runs_api_pipeline_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pipeline Runs"
      },
      "post": {
        "operationId": "start_pipeline_run_api_pipeline_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Pipeline Run"
      }
    },
    "/api/pipeline-runs/{run_id}": {
      "get": {
        "operationId": "get_pipeline_run_api_pipeline_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pipeline Run"
      }
    },
    "/api/pipeline-runs/{run_id}/events": {
      "get": {
        "operationId": "stream_pipeline_events_api_pipeline_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Pipeline Events"
      }
    },
...
  }
}